Move bottom-right buttons to subfolder
