+++ src/buttons/Share.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import RoundButton from './RoundButton';
-import { GetURL } from './url_util';
-import ShareSvg from './svg/share.svg'
+import { GetURL } from '../url_util';
+import ShareSvg from '../svg/share.svg'
 
 const Share = () =>
   <RoundButton onClick={() => navigator.clipboard.writeText(GetURL())}
-               icon={ShareSvg}
-               title='copy url to clipboard'/>
+    icon={ShareSvg}
+    title='copy url to clipboard' />
 
 export default Share;
